requirements.txt
mysql-connector-python
pulp
pandas
faker
numpy
